20 float to double

+++ srcs/ray/hit/hit_cylinder.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   hit_cylinder.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 16:34:42 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/22 16:38:49 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 17:11:49 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-bool	hit_cylinder(t_ray ray, t_cylinder *cylinder, float *t)
+bool	hit_cylinder(t_ray ray, t_cylinder *cylinder, double *t)
 {
 	bool	hit;
-	float	t_body;
-	float	t_caps;
+	double	t_body;
+	double	t_caps;
 
 	hit = false;
 	if (hit_cylinder_body(ray, cylinder, &t_body))
